requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0

# Data Processing
scipy>=1.10.0
h5py>=3.8.0
pillow>=9.5.0

# Training & Evaluation
tqdm>=4.65.0
wandb>=0.15.0

# Visualization
matplotlib>=3.7.0
